alfred_openapi/requirements.txt
apispec ~= 0.27.0
marshmallow ~= 3.0.0b4
pycontracts ~= 1.8.0
